requirements.txt
imagehash
pillow
pillow-heif
docopt
